requirements.txt
boto3
Flask
line_bot_sdk
openai
opensearch_py
python-dotenv
Requests
requests_aws4auth
selenium
webdriver_manager